src/requirements.txt
cloudshell-shell-core>=2.3.0,<2.4.0
enum